requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
aiohttp==3.10.7

azure-functions==1.21.0
azure-functions-durable==1.21.0

azure-storage-blob==12.23.1
azure-identity==1.18.0

sqlalchemy==2.0.35
python-dotenv==1.0.1

PyMuPDF=1.24.10